+++ src/FFMpeg/Format/Audio/DefaultAudio.php
@@ -103,4 +103,12 @@ abstract class DefaultAudio extends EventEmitter implements AudioInterface, Prog
 
         return array($listener);
     }
+
+    /**
+     * {@inheritDoc}
+     */
+    public function getPasses()
+    {
+        return 1;
+    }
 }