+++ src/FFMpeg/Format/Video/DefaultVideo.php
@@ -87,14 +87,6 @@ abstract class DefaultVideo extends DefaultAudio implements VideoInterface
     }
 
     /**
-     * {@inheritDoc}
-     */
-    public function getPasses()
-    {
-        return 1;
-    }
-
-    /**
      * @return integer
      */
     public function getModulus()