+++ tests/FFMpeg/Tests/Media/AudioProg.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace FFMpeg\Tests\Media;
+
+use FFMpeg\Format\ProgressableInterface;
+use FFMpeg\Format\AudioInterface;
+
+abstract class AudioProg implements ProgressableInterface, AudioInterface
+{
+}